Adds test for special collections links when passed to PsuCollectionsLinks
